added a marking if item is already installed and add option to reinstall textures

+++ blender/LilySurfaceScrapper/Scrappers/Cc0texturesScrapper.py
@@ -27,9 +27,11 @@ import re
 from urllib.parse import urlparse, parse_qs, urlencode, urljoin
 from .AbstractScrapper import AbstractScrapper
 
+
 class Cc0texturesScrapper(AbstractScrapper):
     source_name = "CC0 Textures"
     home_url = "https://cc0textures.com"
+    home_dir = "CC0Textures"
 
     @classmethod
     def canHandleUrl(cls, url):
@@ -41,28 +43,28 @@ class Cc0texturesScrapper(AbstractScrapper):
         """Get a list of available variants.
         The list may be empty, and must be None in case of error."""
 
-        if re.match("rcc0\.link/a/", url) is None:
+        if re.match(r"cc0\.link/a/", url) is None:
             query = parse_qs(urlparse(url).query)
             asset_id = query.get('id', query.get('tex', [None]))[0]
         else:
             asset_id = url.split("cc0.link/a/")[-1]
 
         api_url = f"https://cc0textures.com/api/v1/full_json?id={asset_id}"
-        
+
         data = self.fetchJson(api_url)
         if data is None:
             return None
-        
+
         variants_data = data["Assets"][asset_id]["Downloads"]
         variants = list(variants_data.keys())
-        variants_urls = [ variants_data[v]["RawDownloadLink"] for v in variants ]
+        variants_urls = [variants_data[v]["RawDownloadLink"] for v in variants]
 
         self._variants_urls = variants_urls
         self._variants = variants
         self._base_name = asset_id
         return variants
 
-    def fetchVariant(self, variant_index, material_data):
+    def fetchVariant(self, variant_index, material_data, reinstall=False):
         """Fill material_data with data from the selected variant.
         Must fill material_data.name and material_data.maps.
         Return a boolean status, and fill self.error to add error messages."""
@@ -73,20 +75,27 @@ class Cc0texturesScrapper(AbstractScrapper):
         if variant_index < 0 or variant_index >= len(variants):
             self.error = "Invalid variant index: {}".format(variant_index)
             return False
-        
+
         variant = variants[variant_index]
         zip_url = variants_urls[variant_index]
 
-        material_data.name = "CC0Textures/" + self._base_name + "/" + variant
-        zip_path = self.fetchZip(zip_url, material_data.name, "textures.zip")
-        zip_dir = os.path.dirname(zip_path)
-        namelist = []
-        with zipfile.ZipFile(zip_path,"r") as zip_ref:
-            namelist = zip_ref.namelist()
-            zip_ref.extractall(zip_dir)
+        material_data.name = os.path.join(self.home_dir, self._base_name, variant)
 
-        os.remove(zip_path)
-        
+        if reinstall or not self.isDownloaded(variant):
+            zip_path = self.fetchZip(zip_url, material_data.name, "textures.zip")
+            zip_dir = os.path.dirname(zip_path)
+            with zipfile.ZipFile(zip_path, "r") as zip_ref:
+                namelist = zip_ref.namelist()
+                zip_ref.extractall(zip_dir)
+
+            os.remove(zip_path)
+
+            if self.savedVariants is not None:
+                self.savedVariants[variant] = True
+        else:
+            zip_dir = self.getTextureDirectory(material_data.name)
+            namelist = os.listdir(zip_dir)
+
         # Translate cc0textures map names into our internal map names
         maps_tr = {
             # Names of the old website
@@ -113,3 +122,11 @@ class Cc0texturesScrapper(AbstractScrapper):
                 map_name = maps_tr[map_type]
                 material_data.maps[map_name] = os.path.join(zip_dir, name)
         return True
+
+    def isDownloaded(self, variantName):
+        if self.savedVariants is None:
+            self.savedVariants = {i: False for i in self._variants}
+            for i in os.listdir(self.getTextureDirectory(os.path.join(self.home_dir, self._base_name))):
+                self.savedVariants[i] = True
+
+        return self.savedVariants[variantName]
